refactor(tui): extract command runtime handlers component
